ASP-20: Add new fields for communicating per-component pay rates in RP14a
The RP14a schema includes information about a claimant's rate of pay. In
UKSI 2018 No. 1378 a new requirement was added to collect the average rate
of pay following a 52-week reference period for claimants who work variable
hours or for variable rates of pay, for use in calculating holiday pay related
awards.

This change allows multiple specific components of the pay award to have
their rate of pay specified directly, to allow for the 52 week rate to be
supplied for these claimants.