ASP-20: Add `Data to be requested from IP` component rate status
This change adds a new status for component rate to communicate that the
data is known, but cannot be communicated as part of the RP14A upload.
This will be used by the RPS team at Insolvency to initiate direct contact
with the IP.

As with other values, this will only be valid under certain circumstances,
as advised by INSS. At the time of writing, this will be used for HPT claims
for claimants on a variable rate of pay.